Port account_check_printing_aeroo
Because of the increased complexity in version 13.0/14.0 of Odoo,
the enhanced check stub available in aeroo v12.0 was dropped.

This feature could be developped for v14.0.
However, it would require extra time.

For now, the method _check_get_pages() of account.payment can be used to
display stub lines.

See the demo template for an example of usage.